pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-ing"
version = "0.1.0"
description = "A comprehensive RAG application with multiple connectors and dynamic model selection"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-openai>=0.0.1",
    "streamlit>=1.28.0",
    "snowflake-connector-python>=3.0.0",
    "snowflake-sqlalchemy>=1.5.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "tiktoken>=0.5.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "chromadb>=0.4.0",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/selva-k-r/RAG-ing"
Repository = "https://github.com/selva-k-r/RAG-ing"

[project.scripts]
rag-ing = "rag_ing.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true